Raptor_chat - Copy/requirements.txt
flask>=2.0.0
pycryptodome>=3.20.0
python-dotenv>=0.19.0